Moving join to isolated page, recreated layout
